requirements.txt
asgiref==3.2.7
better-exceptions==0.2.2
boto3==1.13.6
botocore==1.16.6
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
dj-database-url==0.5.0
Django==3.0.6
django-debug-toolbar==2.2
django-heroku==0.3.1
django-storages==1.9.1
docutils==0.15.2
folium
geopandas
gunicorn==20.0.4
idna==2.9
Jinja2==2.11.2
jmespath==0.9.5
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
newsapi-python==0.2.6
numpy==1.18.4
pandas==1.0.3
psycopg2==2.8.5
psycopg2-binary==2.8.5
pycountry==19.8.18
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
s3transfer==0.3.3
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.0.1
